request_handlers plugin reorganized
All the code dedicated to the request_handlers plugin has moved
into the specific Python package used for this plugin.
New plugin_helpers package created to store the exceptions that
can be shared among plugins.
