+++ wazo_sysconfd/plugin_helpers/exceptions.py
@@ -1,0 +1,8 @@
+# Copyright 2022 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+
+class HttpReqError(Exception):
+    def __init__(self, code: int, message: str = None):
+        self.code = code
+        self.message = message
