+++ wazo_sysconfd/plugins/request_handlers/chown_autoprov_config.py
@@ -6,7 +6,7 @@ import logging
 
 from pwd import getpwnam
 
-from wazo_sysconfd.request_handlers.command import SimpleCommandFactory
+from wazo_sysconfd.plugins.request_handlers.command import SimpleCommandFactory
 
 logger = logging.getLogger(__name__)
