+++ wazo_sysconfd/plugins/request_handlers/tests/test_asterisk.py
@@ -7,7 +7,7 @@ import unittest
 from mock import ANY, Mock, patch, sentinel
 
 from xivo_bus import Publisher
-from wazo_sysconfd.request_handlers.asterisk import (
+from wazo_sysconfd.plugins.request_handlers.asterisk import (
     AsteriskCommandFactory,
     AsteriskCommandExecutor,
 )
